FIX: Make backends shuffle deterministic
I believe there is a subtle bug in this code. For this algorithm to work, backends[] needs to be equivalent for each invocation of a client in the same round. Just passing backends to random.shuffle is not enough, since that will just deterministically shuffle the array. For the output array to be equivalent the input array must be the same, which is not the case in the current implementation. Getting around that by resetting backends back to the original state (with a sort) before each shuffle. Before this fix it is possible for clients in the same round to route to the  same backend, something that does not happen after this change.

client 0 -> backnd [0, 6, 3]
client 1 -> backnd [5, 1, 7]
client 2 -> backnd [11, 9, 2]
client 3 -> backnd [4, 8, 10]
client 4 -> backnd [8, 11, 4]
client 5 -> backnd [0, 5, 6]
client 6 -> backnd [10, 3, 2]
client 7 -> backnd [7, 9, 1]